This commit ensures that the newly generated psuedo genomes match the well-loved index provided by the ref genome.
Hence if there is an insertion into the psuedo genome, it is removed to keep the index "inframe", however the base before is masked via an N. To prevent primers from being picked across the region.

Getting all this to report into the log file took alot of effort. However, any changes done to the genomes are not recorded under the log file.